+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+# vTBD
+- [BP-1429](https://movai.atlassian.net/browse/BP-1429): Enable tests using docker compose in dal
+
 # v3.2.8
 - [BP-1390](https://movai.atlassian.net/browse/BP-1434): Caching of flow's node dependencies.
+
 # v3.2.7
 - [BP-1434](https://movai.atlassian.net/browse/BP-1434): Duplicated logs on spawner